+++ lib/mongoid/userstamp/config/user_config.rb
@@ -1,17 +1,16 @@
-# -*- encoding : utf-8 -*-
+# frozen_string_literal: true
 
 module Mongoid
-module Userstamp
+  module Userstamp
+    class UserConfig
+      def initialize(options = {})
+        @controller_current_user = options.delete(:controller_current_user)
+        raise ArgumentError.new("Invalid keys found: #{options.keys.join(', ')}") unless options.empty?
+      end
 
-  class UserConfig
-
-    def initialize(opts = {})
-      @reader = opts.delete(:reader)
-    end
-
-    def reader
-      @reader || Mongoid::Userstamp.config.user_reader
+      def controller_current_user
+        @controller_current_user || Mongoid::Userstamp.config.controller_current_user
+      end
     end
   end
-end
 end