+++ lib/mongoid_userstamp.rb
@@ -1,8 +1,8 @@
-# -*- encoding : utf-8 -*-
+# frozen_string_literal: true
 
 require 'mongoid/userstamp'
 require 'mongoid/userstamp/version'
-require 'mongoid/userstamp/config/gem_config'
+require 'mongoid/userstamp/config/global_config'
 require 'mongoid/userstamp/config/model_config'
 require 'mongoid/userstamp/config/user_config'
 require 'mongoid/userstamp/mixins/user'