+++ spec/spec_helper.rb
@@ -1,16 +1,10 @@
-# -*- encoding : utf-8 -*-
-require 'rubygems'
-require 'ostruct'
+# frozen_string_literal: true
 
-$:.push File.expand_path('../../lib', __FILE__)
+$LOAD_PATH.push File.expand_path('../lib', __dir__)
 
-require 'active_support/all'
 require 'mongoid'
 require 'mongoid_userstamp'
-
-%w(admin user book post).each do |file_name|
-  require "support/#{file_name}"
-end
+Dir.glob('support/**/*.rb', base: __dir__).sort.each {|f| require(f) }
 
 Mongoid.configure do |config|
   config.connect_to 'mongoid_userstamp_test'
@@ -18,7 +12,7 @@ end
 
 RSpec.configure do |config|
   config.mock_with :rspec
-  
+
   config.after :suite do
     Mongoid.purge!
   end