+++ spec/support/post.rb
@@ -1,11 +1,12 @@
-# -*- encoding : utf-8 -*-
+# frozen_string_literal: true
+
 class Post
   include Mongoid::Document
   include Mongoid::Userstamp
 
-  mongoid_userstamp user_model: 'Admin',
-                    created_name: :writer,
-                    updated_name: :editor
+  userstamp user_class_name: 'Admin',
+            created_by_field: :writer,
+            updated_by_field: :editor
 
   field :title
 end