+++ spec/unit/model_config_spec.rb
@@ -1,31 +1,34 @@
-# -*- encoding : utf-8 -*-
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe Mongoid::Userstamp::ModelConfig do
-
   subject { Mongoid::Userstamp::ModelConfig.new }
-  before  { Mongoid::Userstamp.stub('config').and_return(OpenStruct.new(created_name: :created_by,
-                                                                        updated_name: :updated_by)) }
-  before  { Mongoid::Userstamp.stub('user_classes').and_return(['User']) }
+  before  do
+    Mongoid::Userstamp.stub('config').and_return(OpenStruct.new(created_by_field: :created_by,
+                                                                updated_by_field: :updated_by))
+  end
+  before { Mongoid::Userstamp.stub('user_classes').and_return(['User']) }
 
   describe '#initialize' do
-
     context 'with opts hash' do
-      subject { Mongoid::Userstamp::ModelConfig.new(user_model: :bar,
-                                                    created_name: :c_by,
-                                                    updated_name: :u_by) }
+      subject do
+        Mongoid::Userstamp::ModelConfig.new(user_class_name: 'Bar',
+                                            created_by_field: :c_by,
+                                            updated_by_field: :u_by)
+      end
 
-      it { should be_a Mongoid::Userstamp::ModelConfig }
-      it { subject.user_model.should eq :bar }
-      it { subject.created_name.should eq :c_by }
-      it { subject.updated_name.should eq :u_by }
+      it { is_expected.to be_a Mongoid::Userstamp::ModelConfig }
+      it { expect(subject.user_class_name).to eq 'Bar' }
+      it { expect(subject.created_by_field).to eq :c_by }
+      it { expect(subject.updated_by_field).to eq :u_by }
     end
 
     context 'without opts hash' do
-      it { should be_a Mongoid::Userstamp::ModelConfig }
-      it { subject.user_model.should eq 'User' }
-      it { subject.created_name.should eq :created_by }
-      it { subject.updated_name.should eq :updated_by }
+      it { is_expected.to be_a Mongoid::Userstamp::ModelConfig }
+      it { expect(subject.user_class_name).to eq 'User' }
+      it { expect(subject.created_by_field).to eq :created_by }
+      it { expect(subject.updated_by_field).to eq :updated_by }
     end
   end
 end